+++ examples/default/ios/InstabugTests/InstabugSampleTests.m
@@ -625,4 +625,13 @@
      OCMVerify([mock setAutoMaskScreenshots:IBGAutoMaskScreenshotOptionLabels | IBGAutoMaskScreenshotOptionTextInputs | IBGAutoMaskScreenshotOptionMedia | IBGAutoMaskScreenshotOptionMaskNothing]);
 }
 
+- (void)testSetNetworkLogBodyEnabled {
+    id mock = OCMClassMock([IBGNetworkLogger class]);
+    BOOL isEnabled = YES;
+
+    OCMStub([mock setLogBodyEnabled:isEnabled]);
+    [self.instabugBridge setNetworkLogBodyEnabled:isEnabled];
+    OCMVerify([mock setLogBodyEnabled:isEnabled]);
+}
+
 @end